refactor(playwright): Move element snapshot functions to @cronn/element-snapshot
