+++ .changeset/cyan-bushes-tickle.md
@@ -1,0 +1,6 @@
+---
+"@cronn/playwright-file-snapshots": minor
+"@cronn/element-snapshot": minor
+---
+
+Breaking change: Move element snapshot functions to `@cronn/element-snapshot`
