+++ packages/element-snapshot/src/browser-lib.ts
@@ -1,0 +1,9 @@
+import { snapshotElement } from "./browser/element";
+
+declare global {
+  interface Window {
+    snapshotElement: typeof snapshotElement;
+  }
+}
+
+window.snapshotElement = snapshotElement;
