+++ packages/element-snapshot/src/browser/attribute.ts
