+++ packages/element-snapshot/src/browser/button.ts
