+++ packages/element-snapshot/src/browser/combobox.ts
