+++ packages/element-snapshot/src/browser/container.ts
