+++ packages/element-snapshot/src/browser/dialog.ts
