+++ packages/element-snapshot/src/browser/element.ts
