+++ packages/element-snapshot/src/browser/group.ts
