+++ packages/element-snapshot/src/browser/heading.ts
