+++ packages/element-snapshot/src/browser/input.ts
