+++ packages/element-snapshot/src/browser/link.ts
