+++ packages/element-snapshot/src/browser/progressbar.ts
