+++ packages/element-snapshot/src/browser/role.ts
