+++ packages/element-snapshot/src/browser/selector.ts
