+++ packages/element-snapshot/src/browser/state.ts
