+++ packages/element-snapshot/src/browser/tab.ts
