+++ packages/element-snapshot/src/browser/table.ts
