+++ packages/element-snapshot/src/browser/text.ts
