+++ packages/element-snapshot/src/browser/types.ts
