+++ packages/element-snapshot/src/browser/utils.ts
