--- packages/element-snapshot/src/types.ts
@@ -1,20 +1,0 @@
-export type {
-  ElementRole,
-  ElementSnapshot,
-  NodeRole,
-  NodeSnapshot,
-} from "./snapshots/types";
-export type { TextSnapshot } from "./snapshots/text";
-
-export type { ButtonSnapshot } from "./snapshots/button";
-export type { ComboboxSnapshot, OptionSnapshot } from "./snapshots/combobox";
-export type { ContainerSnapshot } from "./snapshots/container";
-export type { DialogSnapshot } from "./snapshots/dialog";
-export type { GroupSnapshot, RadiogroupSnapshot } from "./snapshots/group";
-export type { HeadingSnapshot } from "./snapshots/heading";
-export type { InputSnapshot } from "./snapshots/input";
-export type { LinkSnapshot } from "./snapshots/link";
-export type { MenuitemSnapshot } from "./snapshots/list";
-export type { ProgressbarSnapshot } from "./snapshots/progressbar";
-export type { TabSnapshot } from "./snapshots/tab";
-export type { ColumnheaderSnapshot } from "./snapshots/table";
